Tester.tcl allows a developer to quickly develop slicer code. It allows changes to be easily made without re-starting the slicer.
Modules, as well as arbitrary files, can easily be re-sourced. Old GUI's
can be thrown away and new ones created.  I find this module cuts down my
development time drastically.

Note that since this module is for developers only, it might make sense
to have this module loaded only if the user asks for it. That is, most
users won't want to see this.