Fixed the main bugs with model hierarchies:
Removed bad logic that allowed only MRML scenes where all models were
in hierarchies or where there were no hierarchies.  Now they can be
mixed, which means models can be created even if hierarchy exists,
etc.

added MainModelGroupsRefreshGUI and MainModelsRefreshGUI to update the
GUI for each model.  Called from Models.tcl (the module) since GUIs
should only exist if the module has been loaded (this is what makes
the GUI).

Added comments to some procedures.

Bugs that still exist:

No way to add a new model into the hierarchy (all models should be
displayed in Hierarchy.tcl).

In Data.tcl an invalid MRML tree can be created by dragging elements
of the hierarchy.

In Data.tcl a begin hierarchy node can be deleted, this does not
remove the hierarchy, just leaves the end node dangling.

Model refs are not written when equal to default when MRML is saved,
so any model not in a hierarchy has no ModelRefID.  This could be
problematic in future (it at least looks really weird in the file).

This is not well tested yet, perhaps there are other bugs.