Hi All-
I checked in some changes to vtkMrmlVolumeNode.cxx, in the calculation of
the RasToIjk, RasToVtk and Position matrices.  The new code should agree
with the old in the sense that the matrices will usually be the same as
they have been, to a few decimals.  However, the new code has some
advantages:

1) You can load a single slice (e.g. an MRT realtime image) as a volume
and display it in its proper position in 3d.  It can be oblique, though to
see it you may have to turn off interpolation.

2) When using oblique volumes, Slicer won't just default to axial when in
slice mode, it figures out which of axial, coronal and sagittal it is
closest to.

3) Gantry tilt works, I think, even for manual header.  Does anybody care?

4) General code simplification.  A number of subroutines have been
removed.  The method used to find the transforms has been streamlined.
See below if you are interested.

If there are problems, please let me know.

Grazie mille,

-Steve

  // The strategy here is as follows.  We create a 4x4 matrix called Ras,
  // which has in its columns the homogeneous RAS coordinates of four
  // corners of the data volume.  We then create another matrix called
  // Ijk, which has in its columns the IJK coordinates of those same four
  // corners.  Then since the conversion matrix RasToIjk satisfies the
  // equation RasToIjk*Ras=Ijk, we find RasToIjk using the formula
  // RasToIjk=Ijk*Inverse(Ras).