Modularization ---
Added a new proc called "" to allow other modules
to build their own slices if need be.  This code was displaced from the proc
MainSlicesBuildVTK. MainSlicesBuildVTK now calls MainSlicesBuildVTKForSliceActor
where the code used to be.

Added new proc called MainSlicesSetAnno which replaces code that was called in
MainSlicesSetOrient and MainSlicesSetOrientAll.