Added a variable called Interactor(activeSlicer) and set it to be Slicer so that "Slicer" is not hardcoded into each procedure making it easier to allow window interactions in the same manner if ever another vtkMrmlSlicer object is added.
This eliminates the need for duplication of these methods or pushing and popping
bindings since setting Interactor(activeSlicer) to myNewSlicer when you
want it to interact with myNewSlicer and then calling the proc
will take care of this.

One would need to set the Interactor(activeSlicer) back to Slicer when they
are done with it (kind of like pushing and popping bindings in this sense).

If a developer ever writes a new method and uses "Slicer" instead of "Interactor(activeSlicer)"
then the method would still work but only on the vtkMrmlSlicer Slicer object.